Build versioning
RE: #45
- Added version to cli and server packages
- Version is injected by make based either on a Env VAR or
autoincremented patch level
- If version is not injected during build we panic
- For testing a simple string is added
- Added unit tests
- Very minor clean up and style improvment
